or1k: switch on DWARF2_ASM_LINE_DEBUG_INFO
The comment that it doesn't work for the assembler stems from
the older assembler in binutils 2.20.1.

Comparing the .debug_line info between the two shows that
what the old assembler generates is indeed erronous.
The new assemblers output looks good from the little I've seen,
so I take a chance here and switch it on.