fix illegal cast to rtx (*insn_gen_fn) (rtx, ...)
The (static arg) generator functions are casted to a var arg
function pointer, making the assumption that the ABI for passing
the arguments will be the same.
On or1k, this isn't the case, var args are always passed on the stack
even if there would be function argument registers still available.

There exists a bugreport for the problem here:
http://gcc.gnu.org/bugzilla/show_bug.cgi?id=12081

and this patch is taken from the suggestion by Rask Ingemann Lambertsen