Treat varargs as stdargs and split arguments between register and stack
This patch changes the ABI of or1k to be able to handle function calls
correctly in which stdargs and varargs are exchanged.
The varargs are treated like stdargs and saved in registers and on the
stack.
The patch uses the method provided by gcc to save the register arguments
on the stack in the prologue of the function. Then the standard
implementation of varargs can be used.
This implementation makes it necessary, that the arguments are splitted
between register and stack.