or1k: cmpxchg_mask: mask set register
The code that emits this instruction will have loaded the value
that should be set in the lower x-bits of the 'set' register and
then left shift the value in the register into the right position.
It is however not guaranteed that the upper bits of the register
is zero, so this will use the mask register (which also have the
the same x-bits set to 1 right shifted into the right position)
to mask in only the relevant bits.

ChangeLog/
	* config/or1k/or1k.md (cmpxchg_mask): Mask set register.