or1k: Remap registers to fix the ABI
The ABI is horrible, in that it has no two call-saved registers
are adjacent.  Which means that RA cannot save DImode or DFmode
values in registers across calls.

Fix this by making GCC's notion of register numbers different
from the hardware and debug notion of register numbers.

In addition, add REG_ALLOC_ORDER.